Retrying checkout deletes dir first

+++ CIScripts/Checkout/Tentacle.ps1
@@ -14,5 +14,9 @@ function Get-TentacleRepos {
         Expand-Archive -Path repos.zip -DestinationPath .
     })
 
-    Move-Item -Force -Path contrail-windows-test -Destination Test
+    # TODO Temporary change, until PR for contrail-vnc putting contrail-windows-test
+    # to directory 'Test', is merged.
+    if (Test-Path contrail-windows-test) {
+        Move-Item -Force -Path contrail-windows-test -Destination Test
+    }
 }
